Extract server_playing_sound
