+++ src/server.cpp
@@ -165,34 +165,6 @@ void *ServerThread::run()
 	return nullptr;
 }
 
-v3f ServerPlayingSound::getPos(ServerEnvironment *env, bool *pos_exists) const
-{
-	if (pos_exists)
-		*pos_exists = false;
-
-	switch (type ){
-	case SoundLocation::Local:
-		return v3f(0,0,0);
-	case SoundLocation::Position:
-		if (pos_exists)
-			*pos_exists = true;
-		return pos;
-	case SoundLocation::Object:
-		{
-			if (object == 0)
-				return v3f(0,0,0);
-			ServerActiveObject *sao = env->getActiveObject(object);
-			if (!sao)
-				return v3f(0,0,0);
-			if (pos_exists)
-				*pos_exists = true;
-			return sao->getBasePosition();
-		}
-	}
-
-	return v3f(0,0,0);
-}
-
 void Server::ShutdownState::reset()
 {
 	m_timer = 0.0f;
